avoid implicit conversions

+++ src/limiter_ui.cpp
@@ -41,5 +41,5 @@ LimiterUi::~LimiterUi() {
 void LimiterUi::on_new_attenuation(double value) {
   attenuation->set_value(1 - value);
 
-  attenuation_label->set_text(level_to_str(util::linear_to_db(value), 0));
+  attenuation_label->set_text(level_to_str(util::linear_to_db(static_cast<float>(value)), 0));
 }
